Derive Clone for Request
With the current API, it is not possible to save owned values of Request
in a Match implementation, because Match is only given a reference to a
Request, and Request does not implement Clone. If you wanted to store a
list of Requests and iterate over them and make assertions on them after
the fact, this would not be possible unless you could Clone owned values
of Request from their references.