Add QueryParamIsMissingMatcher (#88)
* Add QueryParamMissingMatcher

Sometimes it can be useful to match a request for an explicitly missing
parameter. For example, APIs often have a default behavior if a query
parameter is not passed. A user could therefore want to set up two mocks
- one which should respond to a missing query parameter and another that
  should respond to an exact query parameter. This is tricky currently
as we cannot just create the first mock without the query matcher as
this would also match requests that do contain the query.

This PR adds a matcher that only matches requests that do not contain a
specific query parameter.

* rename method and struct