Add MockGuard::is_satisfied
The intention here is to make it easier to handle situations where we
don't know when to expect the mock to be called.

For example, when testing async message consumers, the tests might
publish a message and need to assert that the message consumer sent an
HTTP request.

Unlike request/response-style HTTP endpoints, the tests don't know when
the message consumer has finished processing.

In these cases, one technique is to wait for an expected condition to
become true, e.g. a mock being called!