Implement From<RangeFull> for Times (#155)
I just noticed this is a bit of an odd one out in the implementation. I
believe RangeFull is the default so a user doesn't have to set it
explicitly. But if they did just out of idle habit then there wasn't a
From conversion.

Not really a big thing but I just figured it would be a mildly nicer DX
if `N..` `..N` and `..` all worked as arguments.